feat: '+' and cancel btn scripts
