+++ src/components/AddTransactionBtn/AddTransactionBtn.jsx
@@ -1,20 +1,21 @@
+import { useDispatch } from 'react-redux';
+import { toggleModalAddTransactionOpen } from 'redux/global/globalSlice';
 import { RoundBtn } from './AddTransactionBtn.styled';
 
-export default function AddTransactionBtn({
-  type,
-  content,
-  onClick = null,
-  disabled = false,
-  hasAccent = false,
-}) {
+export default function AddTransactionBtn() {
+  const dispatch = useDispatch();
+
+  const handleModalAddTransactionOpen = () => {
+    dispatch(toggleModalAddTransactionOpen());
+  };
+
   return (
     <RoundBtn
-      type={type}
-      onClick={onClick}
-      disabled={disabled}
-      hasAccent={hasAccent}
+      type="button"
+      hasAccent={true}
+      onClick={handleModalAddTransactionOpen}
     >
-      {content}
+      +
     </RoundBtn>
   );
 }
